+++ Projects/UnitTests/Source/TestCollections.cpp
@@ -197,34 +197,7 @@ TEST_CASE("Test Collections", "[Collections]")
 			a is [1, "red"], [2, "green"], [3, "blue"]
 			
 			-- Add single element to a
-			add "purple" to a
-
-			)";
-
-		auto script = TestExecuteScript(scriptText);
-		REQUIRE(script);
-		REQUIRE(script->GetVariable("a").IsCollection());
-		auto collection = script->GetVariable("a").GetCollection();
-		REQUIRE(collection);
-		REQUIRE(collection->size() == 4);
-		REQUIRE(collection->find(4) != collection->end());
-		REQUIRE(collection->find(4)->second.GetString() == "purple");
-	}
-
-	SECTION("Test adding collection to existing collection")
-	{
-		static const char * scriptText =
-			u8R"(
-			import core
-
-			-- Create collection using an initialization list of key-value pairs		
-			a is [1, "red"], [2, "green"], [3, "blue"]
-			
-			-- Set variable to one of the collection values
-			b is [4, "purple"]
-
-			-- Add elements in b to a
-			add b to a
+			a[(a size + 1)] is "purple"
 
 			)";
 
@@ -248,7 +221,7 @@ TEST_CASE("Test Collections", "[Collections]")
 			a is [1, "red"], [2, "green"], [3, "blue"]
 			
 			-- Remove element by key
-			remove 2 from a
+			a[2] is null
 
 			)";
 
@@ -261,75 +234,4 @@ TEST_CASE("Test Collections", "[Collections]")
 		REQUIRE(collection->find(2) == collection->end());
 	}
 	
-	SECTION("Test erasing multiple elements from collection")
-	{
-		static const char * scriptText =
-			u8R"(
-			import core
-
-			-- Create collection using an initialization list of key-value pairs		
-			a is [1, "red"], [2, "green"], [3, "blue"]
-			
-			-- Remove elements by key
-			remove (1, 2) from a
-
-			)";
-
-		auto script = TestExecuteScript(scriptText);
-		REQUIRE(script);
-		REQUIRE(script->GetVariable("a").IsCollection());
-		auto collection = script->GetVariable("a").GetCollection();
-		REQUIRE(collection);
-		REQUIRE(collection->size() == 1);
-		REQUIRE(collection->find(1) == collection->end());
-		REQUIRE(collection->find(2) == collection->end());
-	}
-
-	SECTION("Test erasing single element by value from collection")
-	{
-		static const char * scriptText =
-			u8R"(
-			import core
-
-			-- Create collection using an initialization list of key-value pairs		
-			a is [1, "red"], [2, "green"], [3, "blue"]
-			
-			-- Remove element by value
-			remove value "red" from a
-
-			)";
-
-		auto script = TestExecuteScript(scriptText);
-		REQUIRE(script);
-		REQUIRE(script->GetVariable("a").IsCollection());
-		auto collection = script->GetVariable("a").GetCollection();
-		REQUIRE(collection);
-		REQUIRE(collection->size() == 2);
-		REQUIRE(collection->find(1) == collection->end());
-	}
-
-	SECTION("Test erasing multiple elements by value from collection")
-	{
-		static const char * scriptText =
-			u8R"(
-			import core
-
-			-- Create collection using an initialization list of key-value pairs		
-			a is [1, "red"], [2, "green"], [3, "blue"]
-			
-			-- Remove multiple elements by value
-			remove values ("red", "green") from a
-
-			)";
-
-		auto script = TestExecuteScript(scriptText);
-		REQUIRE(script);
-		REQUIRE(script->GetVariable("a").IsCollection());
-		auto collection = script->GetVariable("a").GetCollection();
-		REQUIRE(collection);
-		REQUIRE(collection->size() == 1);
-		REQUIRE(collection->find(1) == collection->end());
-		REQUIRE(collection->find(2) == collection->end());
-	}
-
 }