+++ Source/JxLibCore.cpp
@@ -48,7 +48,7 @@ static Variant WriteLine(ScriptPtr, Parameters params)
 	return nullptr;
 }
 
-static Variant Size(ScriptPtr, Parameters params)
+static Variant GetSize(ScriptPtr, Parameters params)
 {
 	switch (params[0].GetType())
 	{
@@ -80,113 +80,6 @@ static Variant IsEmpty(ScriptPtr, Parameters params)
 	return nullptr;
 }
 
-static void AddToCollection(CollectionPtr collection, Variant value)
-{
-	auto index = collection->size() + 1;
-	while (true)
-	{
-		Variant key = static_cast<int64_t>(index);
-		if (collection->find(key) == collection->end())
-		{
-			collection->insert(std::make_pair(key, value));
-			return;
-		}
-		++index;
-	}
-}
-
-static Variant AddTo(ScriptPtr, Parameters params)
-{
-	if (!params[1].IsCollection())
-		return nullptr;
-
-	// If the first parameter is a collection, then we merge the maps
-	if (params[0].IsCollection())
-	{
-		auto collToInsert = params[0].GetCollection();
-		auto collPtr = params[1].GetCollection();
-		if (collToInsert == collPtr)
-			return nullptr;
-		for (auto itr = collToInsert->begin(); itr != collToInsert->end(); ++itr)
-		{
-			if (collPtr->find(itr->first) != collPtr->end())
-				AddToCollection(collPtr, itr->second);
-			else
-				collPtr->insert(std::make_pair(itr->first, itr->second));
-		}
-		collPtr->insert(collToInsert->begin(), collToInsert->end());
-	}
-	else
-	{
-		AddToCollection(params[1].GetCollection(), params[0]);
-	}
-	return nullptr;
-}
-
-static Variant RemoveFrom(ScriptPtr, Parameters params)
-{
-	if (!params[1].IsCollection())
-		return nullptr;
-
-	// If the first param is a collection, then erase all values, assuming they
-	// are indices in the second parameter.
-	if (params[0].IsCollection())
-	{
-		auto collTarget = params[1].GetCollection();
-		const auto & coll = *params[0].GetCollection();
-		if (collTarget && !collTarget->empty())
-		{
-			for (const auto & v : coll)
-			{
-				collTarget->erase(v.second);
-				if (collTarget->empty())
-					break;
-			}
-		}
-	}
-	else
-	{
-		auto collTarget = params[1].GetCollection();
-		collTarget->erase(params[0]);
-	}
-	return nullptr;
-}
-
-static Variant RemoveValuesFrom(ScriptPtr, Parameters params)
-{
-	if (!params[1].IsCollection())
-		return nullptr;
-
-	if (params[0].IsCollection())
-	{
-		auto & collTarget = *params[1].GetCollection();
-		auto & coll = *params[0].GetCollection();
-		for (const auto & val : coll)
-		{
-			for (auto itr = collTarget.begin(); itr != collTarget.end();)
-			{
-				if (val.second == itr->second)
-					itr = collTarget.erase(itr);
-				else
-					++itr;
-			}
-		}
-	}
-	else
-	{
-		auto & collTarget = *params[1].GetCollection();
-		auto val = params[0];
-		for (auto itr = collTarget.begin(); itr != collTarget.end();)
-		{
-			if (val == itr->second)
-				itr = collTarget.erase(itr);
-			else
-				++itr;
-		}
-	}
-	return nullptr;
-}
-
 void Jinx::RegisterLibCore(RuntimePtr runtime)
 {
 	auto library = runtime->GetLibrary("core");
@@ -194,11 +87,8 @@ void Jinx::RegisterLibCore(RuntimePtr runtime)
 	// Register core functions
 	library->RegisterFunction(true, false, { "write", "{}" }, Write);
 	library->RegisterFunction(true, false, { "write", "line", "{}" }, WriteLine);
-	library->RegisterFunction(true, true, { "{}", "size" }, Size);
-	library->RegisterFunction(true, true, { "{}", "is", "empty" }, IsEmpty);
-	library->RegisterFunction(true, false, { "add", "{}", "to", "{}" }, AddTo);
-	library->RegisterFunction(true, false, { "remove", "{}", "from", "{}" }, RemoveFrom);
-	library->RegisterFunction(true, false, { "remove", "value/values", "{}", "from", "{}" }, RemoveValuesFrom);
+	library->RegisterFunction(true, true, { "{}", "(get)", "size" }, GetSize);
+	library->RegisterFunction(true, true, { "{}", "(is)", "empty" }, IsEmpty);
 
 	// Register core properties
 	library->RegisterProperty(true, true, { "newline" }, "\n");